replaced sidebar with new component and fixed router troubles - sorry dan
